ex01:ok

+++ CPP_42/cpp_06/ex01/Serializer.hpp
@@ -1,11 +1,12 @@
 #pragma once
 #include "Data.hpp"
+#include <stdint.h>
 
 // Use unsigned long instead of uintptr_t
 class Serializer {
 public:
-    static unsigned long serialize(Data* ptr);
-    static Data* deserialize(unsigned long raw);
+    static uintptr_t serialize(Data* ptr);
+    static Data* deserialize(uintptr_t raw);
 private:
     Serializer(); // Prevent instantiation
 };
